+++ docs/news.rst
@@ -12,6 +12,8 @@ Tests:
 
   - Run pure-python ``NameMapper`` tests in a separate process.
 
+  - Fixed a bug in tests with pure-python ``NameMapper``.
+
   - Add Python 3.10 to ``tox.ini``.
 
 3.2.7b1 (2021-07-25)